1.1 and 1.2
Creating the class with some basic "prototype" properties was pretty straight forward and simple. The method is a little more especially when trying to log in. I was having a lot more difficulty by exercise 2.2 with how to correctly use the arrow function. I believe i need more guidance and experience with the simpler items. At times I can't recall how to create a function without looking much less create one as a property of an object and make sure i can still invoke it, or call it outside of where it is.